Decomission of SEB controller
