+++ controllers/starter.go
@@ -113,14 +113,6 @@ func SetupAllReconcilers(mgr controllerruntime.Manager, cfg *config.OperatorConf
 		return err
 	}
 
-	if err = (&SnapshotEnvironmentBindingReconciler{
-		k8sClient:     mgr.GetClient(),
-		Scheme:        mgr.GetScheme(),
-		Configuration: cfg,
-	}).SetupWithManager(mgr); err != nil {
-		return err
-	}
-
 	if err = (&ApplicationReconciler{
 		k8sClient: mgr.GetClient(),
 	}).SetupWithManager(mgr); err != nil {
